+++ .github/workflows/ci.yml
@@ -56,10 +56,6 @@ jobs:
 
       - name: Python tests
         run: |
-          python -m unittest discover -s test -v
-
-      - name: Python tests new
-        run: |
           pytest -vv pymaster
 
       - name: Clean install with threads
@@ -74,4 +70,4 @@ jobs:
 
       - name: Python tests with threads
         run: |
-          python -m unittest discover -s test -v
+          pytest -vv pymaster
